+++ czkawka_cli/src/main.rs
@@ -16,6 +16,7 @@ use czkawka_core::{
     invalid_symlinks::InvalidSymlinks,
     same_music::SameMusic,
     similar_images::{return_similarity_from_similarity_preset, SimilarImages},
+    similar_videos::SimilarVideos,
     temporary::{self, Temporary},
     zeroed::{self, ZeroedFiles},
 };
@@ -384,6 +385,41 @@ fn main() {
             #[cfg(not(debug_assertions))] // This will show too much probably unnecessary data to debug, comment line only if needed
             br.print_results();
             br.get_text_messages().print_messages();
+        }
+        Commands::SimilarVideos {
+            directories,
+            excluded_directories,
+            excluded_items,
+            file_to_save,
+            not_recursive,
+            tolerance,
+            minimal_file_size,
+            maximal_file_size,
+            allowed_extensions,
+        } => {
+            let mut vr = SimilarVideos::new();
+
+            vr.set_included_directory(directories.directories);
+            vr.set_excluded_directory(excluded_directories.excluded_directories);
+            vr.set_excluded_items(excluded_items.excluded_items);
+            vr.set_allowed_extensions(allowed_extensions.allowed_extensions.join(","));
+            vr.set_recursive_search(!not_recursive.not_recursive);
+            vr.set_minimal_file_size(minimal_file_size);
+            vr.set_maximal_file_size(maximal_file_size);
+            vr.set_tolerance(tolerance);
+
+            vr.find_similar_videos(None, None);
+
+            if let Some(file_name) = file_to_save.file_name() {
+                if !vr.save_results_to_file(file_name) {
+                    vr.get_text_messages().print_messages();
+                    process::exit(1);
+                }
+            }
+
+            #[cfg(not(debug_assertions))] // This will show too much probably unnecessary data to debug, comment line only if needed
+            vr.print_results();
+            vr.get_text_messages().print_messages();
         }
         Commands::Tester { test_image } => {
             if test_image {